fibonacci_function/requirements.txt
azure-functions
mpmath==1.3.0